Update modern media controls testing utilities https://bugs.webkit.org/show_bug.cgi?id=172976 <rdar://problem/32588432>
Reviewed by Eric Carlson.

Source/WebCore:

Stop logging an error when we're not able to load UI Strings since UI strings
are not defined in the majority of modern-media-controls tests as we load the
code directly from the modern-media-controls module and can't reach into the
English.lproj directory, nor is it necessary.

* Modules/modern-media-controls/main.js:
(UIString):

LayoutTests:

Update the list of CSS files to load to match the current list of
existing CSS files in the modern-media-controls module.

* media/modern-media-controls/resources/media-controls-loader.js:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@217826 268f45cc-cd09-0410-ab3c-d52691b4dbfc