[GCrypt] RSA-PSS support https://bugs.webkit.org/show_bug.cgi?id=172856
Reviewed by Jiewen Tan.

.:

* Source/cmake/OptionsGTK.cmake: Set and expose the USE_GCRYPT preprocessor macro when
SUBTLE_CRYPTO is enabled.
* Source/cmake/OptionsWPE.cmake: Match the GTK+ port in demanding at least version 1.7.0
of libgcrypt. Also set and expose the USE_GCRYPT preprocessor macro.

Source/WebCore:

Implement RSA-PSS support for Web Crypto API in ports that use libgcrypt.

For the signing operation, the digest algorithm is determined and used to
compute the hash of the passed-in data. The given salt length and hash results
are embedded into a data s-expression, specifying PSS as the preferred type of
padding. This s-expression and the provided key are then passed to the
gcry_pk_sign() call which computes the signature. The MPI data of this signature
is retrieved and its size validated before being passed to the success callback.

For the verification operation, the hash of the passed-in data is computed as
before, embedding it into a data s-expression along with the specified salt
length. The passed-in signature data is embedded into a sig-val s-expression.
The two s-expressions are passed to the gcry_pk_verify() call along with the
key. If the returned error code for this operation is GPG_ERR_NO_ERROR the
verification is deemed successful, while in case of any other error code (which
might indicate a verification failure or an operation error) we simply signal
that the verification step failed (and not that it possibly errored out).

No new tests -- covered by existing tests.

* PlatformGTK.cmake: Add the new build targets.
* PlatformWPE.cmake: Ditto.
* crypto/gcrypt/CryptoAlgorithmRSA_PSSGCrypt.cpp: Added.
(WebCore::gcryptSign):
(WebCore::gcryptVerify):
(WebCore::CryptoAlgorithmRSA_PSS::platformSign):
(WebCore::CryptoAlgorithmRSA_PSS::platformVerify):
* crypto/gcrypt/CryptoAlgorithmRegistryGCrypt.cpp:
(WebCore::CryptoAlgorithmRegistry::platformRegisterAlgorithms): Register the RSA-PSS algorithm.

Source/WTF:

* wtf/Platform.h: Define HAVE_RSA_PSS to 1 for USE(GCRYPT). Support for RSA-PSS
is provided by default through the libgcrypt library.

LayoutTests:

* platform/gtk/TestExpectations: Enable RSA-PSS tests that we pass. The remaining
disabled tests rely on SPKI or PKCS#8 support that hasn't landed yet.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@217877 268f45cc-cd09-0410-ab3c-d52691b4dbfc