CSP: Apply img-src directive to favicon loads https://bugs.webkit.org/show_bug.cgi?id=172733
Remove CONSOLE MESSAGE line from expected result that I inadvertently left in from an earlier iteration
of the test that did not call testRunner.queueReload(). Calling testRunner.queueReload() clears out
any emitted console messages. The dump of the resource load callbacks is sufficient to ensure that
the favicon was blocked as expected.

* http/tests/security/contentSecurityPolicy/block-favicon-expected.txt:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@218026 268f45cc-cd09-0410-ab3c-d52691b4dbfc