Modern media controls tests error in Button.js https://bugs.webkit.org/show_bug.cgi?id=173439
Patch by Antoine Quint <graouts@apple.com> on 2017-06-15
Reviewed by Dean Jackson.

Tests would sometime yield an error when commit() would be called and the _imageSource
ivar hadn't been set. To more safely commit the mask image when it's loaded, we now use
a markDirtyProperty() call and a commitProperty() override to ensure that we cover the
case where we want to commit the mask image, and for other commits not to have to worry
about the mask image being defined.

* Modules/modern-media-controls/controls/button.js:
(Button.prototype.commitProperty):
(Button.prototype._updateImage):
(Button.prototype.commit): Deleted.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@218367 268f45cc-cd09-0410-ab3c-d52691b4dbfc