Meter element doesn't respect the writing direction https://bugs.webkit.org/show_bug.cgi?id=173507
Reviewed by Sam Weinig.

Source/WebCore:

The bug was caused by NSLevelIndicatorCell no longer using the value of baseWritingDirection
to determine the direction of rendering in macOS Sierra and later. It instead relies on
the value of userInterfaceLayoutDirection.

Fixed the bug by setting both values. Once we dropped the support for macOS El Capitan
and earlier, we can remove the code to set baseWritingDirection.

Test: fast/dom/HTMLMeterElement/meter-rtl.html

* rendering/RenderThemeMac.mm:
(WebCore::RenderThemeMac::levelIndicatorFor):

LayoutTests:

A regression ref test since existing tests were pixel tests and did not catch this regression.
Note that black border was added to workaround the anti-aliasing differences in macOS High Sierra.

* fast/dom/HTMLMeterElement/meter-rtl-expected.html: Added.
* fast/dom/HTMLMeterElement/meter-rtl.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@218468 268f45cc-cd09-0410-ab3c-d52691b4dbfc