[iOS] Always include frames in the scrolling tree when async frame scrolling is enabled https://bugs.webkit.org/show_bug.cgi?id=173405
Patch by Frederic Wang <fwang@igalia.com> on 2017-06-19
Reviewed by Simon Fraser.

Source/WebCore:

Currently "async frame scrolling" is ignored on iOS. This commit changes that behavior to
align on macOS and is a preliminary step to implement iframe scrolling on iOS (bug 149264).

Test: compositing/iframes/compositing-for-scrollable-iframe.html
      fast/scrolling/scrolling-tree-includes-frame.html

* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::requiresCompositingForScrollableFrame): Do not require
compositing when "async frame scrolling" is enabled on iOS.

Source/WebKit2:

* WebProcess/WebCoreSupport/WebChromeClient.h: Set the ScrollableNonMainFrameTrigger bit by
default on iOS too. RenderLayerCompositor::requiresCompositingForScrollableFrame will ignore
this change when async frame scrolling is disabled.

LayoutTests:

Add some new expectations for iOS/WK2 now that the "async frame scrolling" setting is taken
into account for the layer and scrolling trees.

* platform/ios-wk2/compositing/iframes/compositing-for-scrollable-iframe-expected.txt: Added.
A specific layer is created for the iframe.
* platform/ios-wk2/fast/scrolling/scrolling-tree-includes-frame-expected.txt: Added.
The frame appears in the scrolling tree.
* platform/ios/fast/scrolling/scrolling-tree-includes-frame-expected.txt: Renamed from LayoutTests/platform/ios-simulator/fast/scrolling/scrolling-tree-includes-frame-expected.txt.
Use this as a generic expectation for iOS.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@218535 268f45cc-cd09-0410-ab3c-d52691b4dbfc