Remove fast/regions/region-leak-js-information-when-disabled-at-runtime.html https://bugs.webkit.org/show_bug.cgi?id=173784
Patch by Frederic Wang <fwang@igalia.com> on 2017-06-23
Reviewed by Simon Fraser.

fast/regions/region-leak-js-information-when-disabled-at-runtime.html is supposed to test
that some JS properties are undefined when CSS regions are disabled at runtime. However, this
part is currently commented out and the test is no longer relevant after r200524. Hence we
just remove it.

* fast/regions/region-leak-js-information-when-disabled-at-runtime-expected.txt: Removed.
* fast/regions/region-leak-js-information-when-disabled-at-runtime.html: Removed.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@218762 268f45cc-cd09-0410-ab3c-d52691b4dbfc