[GCrypt] Properly initialize libgcrypt before using it https://bugs.webkit.org/show_bug.cgi?id=173589
Reviewed by Michael Catanzaro.

Initialize libgcrypt in the Unix-specific main() entrypoint. This is early enough
to ensure no other libgcrypt API function is invoked and to also ensure this is
done in a thread-safe manner.

Initialization is initiated through the gcry_check_version() call. 16 kilobytes of
secure memory is pre-allocated before we mark the initialization as complete, as
recommended by the libgcrypt documentation.

* WebProcess/EntryPoint/unix/WebProcessMain.cpp:
(main):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@218814 268f45cc-cd09-0410-ab3c-d52691b4dbfc