RegExpPrototype.js builtin uses for-of iteration which is almost certainly incorrect https://bugs.webkit.org/show_bug.cgi?id=173740
Reviewed by Mark Lam.

JSTests:

* stress/regexp-prototype-replace-builtin-should-not-use-for-of.js: Added.
(Array.prototype.Symbol.iterator):

Source/JavaScriptCore:

The builtin was using for-of iteration to iterate over an internal
list in its algorithm. For-of iteration is observable via user code
in the global object, so this approach was wrong as it would break if
a user changed the Array iteration protocol in some way.

* builtins/RegExpPrototype.js:
(replace):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@218828 268f45cc-cd09-0410-ab3c-d52691b4dbfc