lint-test-expectations should be run during style checking https://bugs.webkit.org/show_bug.cgi?id=173559 <rdar://problem/32854941>
Reviewed by Aakash Jain.

* Scripts/webkitpy/layout_tests/models/test_expectations.py:
(TestExpectationsModel._already_seen_better_match): Append returns 'None' when successful,
prevents identifying errors on specific lines in a file.
* Scripts/webkitpy/style/main_unittest.py:
(ExpectationLinterInStyleCheckerTest.test_linter_duplicate_line_edit_in_file): Test for the
case where a linter error is in the file modified, but the line modified is not related to
the linter error.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219689 268f45cc-cd09-0410-ab3c-d52691b4dbfc