Turn tests at media/modern-media-controls/start-support back on https://bugs.webkit.org/show_bug.cgi?id=174683
Patch by Antoine Quint <graouts@apple.com> on 2017-07-20
Reviewed by Dean Jackson.

Source/WebCore:

Turning those tests back on revealed a small bug that is unlikely to really affect content
on the Web. In the case where the size of the video is known right away, without the need
for loading its metadata, as is the case in the start-support-click-to-start.html test with
a local media resource, all queued layouts are flushed at once and we may call the layout()
method of the left ButtonsContainer which originally is set to contain the play/pause button,
which would remove the play/pause button from the center of the media. So before we potentially
set the play/pause button as the central button, we first assign the default button set for
the two ButtonsContainer instances and only add the play/pause button when we're not showing
the prominent play/pause button.

* Modules/modern-media-controls/controls/inline-media-controls.js:
(InlineMediaControls.prototype.layout):
(InlineMediaControls.prototype._leftContainerButtons):

LayoutTests:

* media/modern-media-controls/start-support/start-support-click-to-start-expected.txt:
* media/modern-media-controls/start-support/start-support-click-to-start.html:
* media/modern-media-controls/start-support/start-support-fullscreen.html:
* media/modern-media-controls/start-support/start-support-lowPowerMode-expected.txt:
* media/modern-media-controls/start-support/start-support-lowPowerMode.html:
* platform/ios-simulator/TestExpectations:
* platform/mac/TestExpectations:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219700 268f45cc-cd09-0410-ab3c-d52691b4dbfc