[GTK] Icon database error and crash https://bugs.webkit.org/show_bug.cgi?id=174760
Reviewed by Michael Catanzaro.

The crash is a debug ASSERT that happens when the IconRecord image is created in one thread and destroyed in
another one. IconDatabase creates and destroys IconRecord objects in both database and main thread. The
IconRecord is destroyed when the icon is no longer retained, and we only release icons when we fail to get the
image data (including pages that don't have a favicon). We can prevent this crash from happening if we ensure we
never create an Image for an IconRecord when the given image data is nullptr.

* UIProcess/API/glib/IconDatabase.cpp:
(WebKit::IconDatabase::IconRecord::setImageData):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219861 268f45cc-cd09-0410-ab3c-d52691b4dbfc