REGRESSION(r218629): [GTK] ASSERTION FAILED: m_eglDisplay == EGL_NO_DISPLAY in WebCore::PlatformDisplay::~PlatformDisplay https://bugs.webkit.org/show_bug.cgi?id=174789
Reviewed by Žan Doberšek.

Before r218629, the wayland nested compositor display was a singleton created on demand by
AcceleratedSurfaceWayland, but now it's created unconditionally on web process creation when a valid display
name is received from the UI process. The side effect of this is that now the shared display is created after
the nested compositor display that is not a singleton anymore. The atexit used by PlatformDisplay to shutdown
egl display assumes that it will be registered by the shared display. Since we switched to on demand AC mode,
the shared display is only used in the web process when entering AC mode, that's why it only crashes when we
have entered AC at least once. We just need to ensure that the shared display is created before the nested
compositor display.

* WebProcess/gtk/WaylandCompositorDisplay.cpp:
(WebKit::WaylandCompositorDisplay::create): Check we are under wayland before creating the WaylandCompositorDisplay.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219866 268f45cc-cd09-0410-ab3c-d52691b4dbfc