MediaDocument fullscreen pinch gesture should trigger navigate back. https://bugs.webkit.org/show_bug.cgi?id=174914
Patch by Jeremy Jones <jeremyj@apple.com> on 2017-07-27
Reviewed by Jon Lee.

No new tests because only effect is from interaction with platform.

For media documents, pressing the "done" button on fullscreen video navigates back
to the previous page. The same should happen for other gestures that pause playback
when returning to inline. This allows the gesture to have the same behavior as the
button.

* platform/ios/WebVideoFullscreenInterfaceAVKit.mm:
(WebVideoFullscreenInterfaceAVKit::shouldExitFullscreenWithReason):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219991 268f45cc-cd09-0410-ab3c-d52691b4dbfc