JSContext Inspector: Scripts sometimes do not show in resources tab https://bugs.webkit.org/show_bug.cgi?id=175153 <rdar://problem/33708683>
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2017-08-03
Reviewed by Matt Baker.

* UserInterface/Views/ResourceSidebarPanel.js:
(WI.ResourceSidebarPanel.prototype.initialLayout):
When the ResourceSidebar is lazily created, be sure to add any scripts
to the sidebar that are not backed by Resources.

(WI.ResourceSidebarPanel.prototype._scriptWasAdded):
(WI.ResourceSidebarPanel.prototype._addScript):
Extract so it can be used outside of an event handler.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@220251 268f45cc-cd09-0410-ab3c-d52691b4dbfc