[GTK][WPE] Assertion failure in TimerBase inside WebCore::IconRecord::setImageData https://bugs.webkit.org/show_bug.cgi?id=173866 <rdar://problem/33122050>
Reviewed by Michael Catanzaro.

IconDatabase creates and destroys IconRecord objects in both database and main thread. If the IconRecord has a
valid icon, its Image could be created in one thread and destroyed in another, something that is not expected to
happen, because Image has a Timer member. Since we have all the data and we are decoding it right after creating
the Image, we don't really need to keep the Image object around, we could simply take a reference of the encoded
data and the decoded native image to be returned by synchronousIconForPageURL().

* UIProcess/API/glib/IconDatabase.cpp:
(WebKit::IconDatabase::IconRecord::image): Return NativeImagePtr now.
(WebKit::IconDatabase::IconRecord::setImageData): Create a BitmapImage to decode it and keep a reference to the
encoded data and decoded native image.
(WebKit::IconDatabase::IconRecord::snapshot const): Use m_imageData to get the encoded data.
(WebKit::IconDatabase::synchronousIconForPageURL): Return the native image and whether the icon is known or not.
(WebKit::IconDatabase::IconRecord::loadImageFromResource): Deleted.
* UIProcess/API/glib/IconDatabase.h:
* UIProcess/API/glib/WebKitFaviconDatabase.cpp:
(getIconSurfaceSynchronously): Use new API.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@220580 268f45cc-cd09-0410-ab3c-d52691b4dbfc