REGRESSION(r220517-r220521) [GTK] Various compositing tests fail https://bugs.webkit.org/show_bug.cgi?id=175478
Reviewed by Sergio Villar Senin.

Source/WebCore:

Fix a regression introduced in r220519. Move a framebuffer binding in TextureMapperGL
away from GraphicsContext3D and onto the OpenGL API to work around the framebuffer
binding state that's kept in the GraphicsContext3D class. This mismatch in how the
framebuffer was bound was causing a couple of test failures in compositing tests.

No new tests -- regressed tests have their failure expectations removed.

* platform/graphics/texmap/TextureMapperGL.cpp:
(WebCore::TextureMapperGL::bindDefaultSurface): Don't use GraphicsContext3D for the
glBindFramebuffer() call. This would be done in subsequent patches that would move
TextureMapperGL off the GraphicsContext3D usage, but this specific use case clashes
with the framebuffer binding that's done in BitmapTextureGL and was causing test
failures.

LayoutTests:

* platform/gtk/TestExpectations: Remove failure and flakiness expectations.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@220699 268f45cc-cd09-0410-ab3c-d52691b4dbfc