Only create directory and sandbox extension handle for storage directories if they are set https://bugs.webkit.org/show_bug.cgi?id=175522
Reviewed by Brady Eidson.

Source/WebKit:

* UIProcess/WebsiteData/WebsiteDataStore.cpp:
(WebKit::WebsiteDataStore::resolveDirectoriesIfNecessary):
If using a non-persistent WebsiteDataStore, all the storage directories
are unset, but we still try to create them and make sandbox extensions
for them. This has no effect except a ton of noisy logging, but that
is reason enough to stop.

This is similar to r216053, but for more directories.

Tools:

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/WebKit2Cocoa/WKWebViewDoesNotLogDuringInitialization.mm: Added.
(TEST):
Add a test that making a WKWebView with an ephemeral WKWebsiteDataStore
doesn't log anything to stderr.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@220709 268f45cc-cd09-0410-ab3c-d52691b4dbfc