Re-enable tests based on media-fullscreen.js for GTK. https://bugs.webkit.org/show_bug.cgi?id=175770
Patch by Ms2ger <Ms2ger@gmail.com> on 2017-08-21
Reviewed by Eric Carlson.

The media-fullscreen-not-in-document test checks that a disconnected
element can go fullscreen, while the code explicitly disallows that
(in Document::requestFullScreenForElement). Update the test to match
the code.

The media-fullscreen-inline test will now fail more clearly if the
attempt to go fullscreen is rejected (which it isn't).

* media/media-fullscreen-not-in-document-expected.txt: rebaseline.
* media/media-fullscreen.js:
(fullscreenerror): pass the test if the webkitfullscreenerror
(addEventListeners): add a webkitfullscreenerror event handler
* platform/gtk/TestExpectations: remove failure expectations.
* platform/gtk/media/media-fullscreen-not-in-document-expected.txt:
  Copied from LayoutTests/media/media-fullscreen-not-in-document-expected.txt.
  This copy leaves out the webkitSupportsPresentationMode check that's
  disabled on GTK.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@220974 268f45cc-cd09-0410-ab3c-d52691b4dbfc