Fix Titzer bench on iOS. https://bugs.webkit.org/show_bug.cgi?id=175917
Reviewed by Ryosuke Niwa.

Currently, Titzer bench doesn't run on iOS since the benchmark
allocates lots of physical pages that it never actually writes
to. We limited the total number wasm physical pages to the ram
size of the phone, which caused us to fail a memory
allocation. This patch changes it so we will allocate up to 3x ram
size, which seems to fix the problem.

* wasm/WasmMemory.cpp:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@221126 268f45cc-cd09-0410-ab3c-d52691b4dbfc