[GTK][Wayland] Use fast malloc to allocate backing store cairo surfaces data https://bugs.webkit.org/show_bug.cgi?id=176226
Reviewed by Michael Catanzaro.

Source/WebCore:

It reduces the memory consumption and improves the performance. Also, since performance is better we don't need
to keep the scroll surface allocated all the time, we can create it on demand and keep it allocated only while
scrolling too fast.

* platform/graphics/cairo/BackingStoreBackendCairoImpl.cpp:
(WebCore::createCairoImageSurfaceWithFastMalloc): Helper to create a cairo image surface using fast malloc.
(WebCore::BackingStoreBackendCairoImpl::BackingStoreBackendCairoImpl): Create the main surface, but no the
scroll one. Initialize the hysteresis activity to discard the scroll surface after scrolling.
(WebCore::BackingStoreBackendCairoImpl::scroll): Create the scroll surface only if it's nullptr, and notify the
hysteresis activity.
* platform/graphics/cairo/BackingStoreBackendCairoImpl.h:

Source/WebKit:

Do not create the cairo surface for BackingStoreBackendCairoImpl. When not using X11 we are going to use an
image surface anyway, so we don't really need to use the create similar. This way we only need to pass the size
and device scale factor to BackingStoreBackendCairoImpl that will create the image surface using fast malloc to
allocate the data.

* UIProcess/cairo/BackingStoreCairo.cpp:
(WebKit::BackingStore::createBackend):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@221521 268f45cc-cd09-0410-ab3c-d52691b4dbfc