[WPE] Use the device scale factor value in WebPageCreationParameters https://bugs.webkit.org/show_bug.cgi?id=175377
Reviewed by Michael Catanzaro.

* WebProcess/WebPage/AcceleratedDrawingArea.cpp:
(WebKit::AcceleratedDrawingArea::AcceleratedDrawingArea):
WebPageCreationParameters::deviceScaleFactor value should be used
to set the device scale factor on the given WebPage object. This
enables correctly leveraging the scale factor for the WPE port
which enters accelerated mode unconditionally and as such doesn't
wait for the backing store update that would provide that value
otherwise.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@221571 268f45cc-cd09-0410-ab3c-d52691b4dbfc