HTMLMediaElement: attach any existing EME CDMInstance in mediaEngineWasUpdated() https://bugs.webkit.org/show_bug.cgi?id=176354
Reviewed by Xabier Rodriguez-Calvar.

It's common for MediaKeys to be set on an HTMLMediaElement object before
any resource is loaded for that element, meaning no underlying MediaPlayer
exists at that point. In order to not stall the pipeline when encrypted
content is finally loaded, the CDMInstance object of the current MediaKeys
set on this element should be attached in mediaEngineWasUpdated() that's
called whenever the underlying MediaPlayerPrivate object is replaced.

* html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::mediaEngineWasUpdated):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@221614 268f45cc-cd09-0410-ab3c-d52691b4dbfc