bisect-builds started failing with invalid json https://bugs.webkit.org/show_bug.cgi?id=176478
Once we hit 1000 items in our dynamoDB, we ran into an upper limit on the AWS API Gateway looping mechanism. 
This prevented us from doing a body mapping template to reformat the json. This fix is to simply return the raw json, 
and adjust the json parsing in our code. The AWS rest API has been fixed, and updated to v2. 
        
This change is for the script which operates against the endpoint.
        
Reviewed by Aakash Jain.

* Scripts/bisect-builds:
(get_sorted_revisions):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@221741 268f45cc-cd09-0410-ab3c-d52691b4dbfc