REGRESSION: [GTK] Generated HTML API documentation is missing descriptions https://bugs.webkit.org/show_bug.cgi?id=176667
Reviewed by Michael Catanzaro.

Source/WebKit:

Fix several typos in API documentation.

* UIProcess/API/glib/WebKitSettings.cpp:
(webkit_settings_class_init):
* UIProcess/API/glib/WebKitWebView.cpp:
(webkit_web_view_class_init):
* UIProcess/API/glib/WebKitWebsiteDataManager.cpp:
* UIProcess/API/gtk/docs/webkit2gtk-docs.sgml:

Tools:

This is happening since we moved the implementation files to glib dirs. The generate-gtkdoc script assumes that
header and implementation files are always in the same directory. To reduce the input we pass to gtk-doc we
ignore cpp files for which there isn't a public header. The code to do this check doesn't consider cpp files in
a different directory than the header file. It also assumes that header and cpp files have the same filename
except for the extension, which is true in all cases but WebKitWebView that we have two files WebKitWebView.cpp
and WebKitWebViewGtk.cpp.

* gtk/generate-gtkdoc:
(files_to_ignore): Check also cpp files with the Gtk suffix.
(files_to_ignore.add_file_if_exists): Check if a file with the given name exists in any of the source dirs.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@221847 268f45cc-cd09-0410-ab3c-d52691b4dbfc